requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
scipy>=1.5.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
Pillow>=8.0.0
